More globals for JSHint
`xit` is for tests - you can pend one this way.

`$`, `Singly` and `Stripe` are all things that I intend to be global, but JSHint
has no way of knowing I've loaded them. I wanted to put the latter two in each
file they're part of, but can't immediately figure out that hint to JSHint.